requirements.txt
contourpy==1.3.1
cycler==0.12.1
fonttools==4.56.0
kiwisolver==1.4.8
matplotlib==3.10.0
numpy==2.2.3
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pyparsing==3.2.1
pyserial==3.5
python-dateutil==2.9.0.post0
python-osc==1.9.3
pytz==2025.1
seaborn==0.13.2
six==1.17.0
supercollider==0.0.6
tzdata==2025.1
